Spu + Memory: added SPU registers and ram mirroring
SPU is the sound processing unit, we need those registers because the
bios is using them, but we will not work on audio until much much later.

Also added KSEG0 and KSEG1 mirroring manually for ram only at this
point, though the approach is not very good, we need to make it better
in the future and include I-cache to it as well.