Cop: moved cop holding to Bus and implemented cache isolation
At some point, the kernel will clear the cache by setting
`cop0.SR = 0x1000`, and then writing to the address `0x00000000`, so we
need to control the RAM access if `cop0.SR == 0x1000`. In order to help
with that, (since in our design the Bus cannot access the Cpu and thus
cannot access `cop0.SR` if it was hold in the Cpu) we need to move it to
the bus as was done in this commit.

Also planning to make the other cops in the bus as well, so made a
specific trait `CpuBusProvider` that will provide `BusLine` of the bus
and a way to control/read/write to coprocessors