Cpu + Memory: moved `cop0` back to the cpu as it is needed in exceptions
Cop0 has all exceptions handling registers and commands, so it will make
it harder for us to use it if it was available behind the bus.

In the first place, we moved `cop0` to the bus in order to filter
reads/writes to the isolated cache (when isolating cache via `cop0.SR`),
so in order to fix this, we build a wrapper read/write functions inside
the CPU, that will forward writes to the bus ram if the cache is not
isolated. If it is isolated, it will ignore read/writes for now, until
we implement I-cache, which can be hold in the CPU as well.