Dma: temporary fix to the DMA timing issue
The goal of this commit is to add a time delay of the DMA2(GPU) in
linked list mode.

The problem was that, one entry in the linked list will change its
`next` (this entry is in address `0x00138CA8`). The bios will put the
new address before running the DMA, then clear it and put a new one.

The problem is in the time of clearing this address, before this commit,
it was clearing it before the DMA reach it and process it, and thus the
DMA reads `0x000000` at that address. The bios does not wait or check if
the DMA is still running. Thus we needed a way to finish quickly, I
tried skipping without adding the `delay` at the end, but it does not
work for some reason (The emulator crashes quickly). So we did this
temporarily.