Gpu: used OpenGL to implement drawing
Used `glium` library to handle drawing using OpenGL.

We have two versions of the VRAM, one in bytes form that we can
read/write to, and Texture form which we draw to and render to the front
window from.
But this separation creates problem, because the two VRAMs have the
color data in different formats, so when we need to convert from one to
another, it will be a heavy process.