Gpu: implemented rendering textured polygons
The textures are being held in `texture_buffer` which is updated from
the VRAM after every transfer to it from the CPU, thus we don't support
texturing from a rendering range yet.

The VRAM now is stored in a Buffer, not sure which is faster in general,
we need to do some testing regarding that.