Cdrom: implemented `0x01` and `0x1A` commands
GetID(0x1A) returns 2 outputs, so I used a trick to be able to handle
that, we can use the same way we did for the GPU here, using
`Box<dyn Command>`, but its an overkill and might slow things, so for
now, we will use this design, until it needs to be changed.

GetID now only returns a fixed value that indicates `No Disk` state,
this is just to get the BIOS completed, and we can then implement things
we need one by one.

This is the last thing we need to implement in the Cdrom to get the BIOS
to display the display that no Disk is present, next we will need to
handle controllers and some more in the GPU.