Gpu + Dma: added support for DMA2 from VRAM to CPU
The dma side is ok, we just flip the stuff. But before this, the GPU
will wait for the clock before getting the next `gpu_read` data ready.
But because we don't clock from the CPU when doing the DMA, we are doing
it manually for now.

Actually, the DMA is working from the CPU, but CPU will not execute
instructions at that time. This is how we should implement the DMA, but
for now, we are doing it in this weird way.

If we run the DMA from the CPU clock (each clock sends one word), then
we don't need to clock the `gp0` commands from `gpu_read` like this.

Needs lots of rewrite later XD