Gpu: implemented texture blending on rendering
Not sure how it works, the docs did not say anything specific about
this.

I assumed value of 0x80 will not change the brightness of the texture,
value of 0xFF will double the brightness and value of 0x00 will result
in black pixels.
If all of the above is true, we can implement that by doing this:

```
texture_pixel *= color_pixel * 2;
```

Thus if the value is 0x80(0.5) it will result in 1.0, 0xFF(1.0) => 2.0, and
0x00(0.0) => 0.0.