Exe Loader: Added Exe loading into ram and patched `run_shell` to run it
Since many testing files come in EXE, which is not CDROM and also as we
don't have support for CDROM yet, adding manual EXE file loading is
better.

I used the method used in `https://github.com/BluestormDNA/ProjectPSX`
to load the exe, which works by loading the exe file into the correct
position in ram, then patching the beginning of the function
`LoadRunShell` in the bios to load the values for PC, GP, SP/FP
registers and jumping to the beginning to the exe file (PC value).

Some emulators use different methods, but this is very simple to
implement and looks like it works well, if we need other methods, we can
add them later.