Gpu: made writing to Vram in blocks and added rendering Vram blits
In some of the testing EXEs, they don't render using polygons and
rectangles and only blit the font characters directly. Now we will be
able to view them.

There is an issue though regarding reading u16 data and converting it
into colors, `glium` library does not support the u1u5u5u5 format at the
time of this commit, but I made a PR to the project to support it, if
they accepted it, the code will be much faster and straight forward.

PR: https://github.com/glium/glium/pull/1943