Cpu: Fixed bug in `Lwr` implementation
The mask was done incorrectly, due to using the same shifts with `lwl`.

The tests in https://github.com/PeterLemon/PSX helped me find this. Now
the CPU passes all the tests in that repository. But the repository does
not test everything, so there might be other bugs here and there.