Github Actions: added CI
This is needed now even before tests because I used code that is not in the commit tree and that resulted in making the app no buildable. The cause is the commit 8be3852a03d8d9223be670e9cfa33074c7361c43. So adding CI should notify me of any similar issues in the future