GTE: added all registers and handled their behaviour or write/read
Some registers update other registers on write, that is handled also in
this commit as well as the addition and the correction of types of all
registers.

The types are not all correct for computation, some types are fractions
(for example 1 bit sign, 3 integer, and 12 fraction for matrix numbers)
but we handled sign extension.