Gpu: moved the gpu `program` to the `GpuContext` for performance
Before, we were compiling `program` on every single draw. It was
temporary, and now we moved it as an attribute in the class to ease
using it.

The next thing I should do is to move the glsl code to separate files.
Since its not very practical to use strings to write code.