Cpu: Added load delay
Load instructions in this CPU takes 2 instructions to complete, meaning
if you load to a register from memory and them read the register (copy
to another register for example) you will get the old value and not the
value that you just read from the memory.

This commits implements this mechanism. There are still some edge cases
(tests are not passing), but this should be good for starters.