Gpu: added code to sync vram data between `drawing_texture` and `vram`
Before this commit, if a program draws into the vram, and then used a
section of the drawing as texture, we will see the content of the vram
which was last synced. Before this commit, only textures copies
manually into the vram work.

But, with all this syncing and copying. It will take a lot of time to
just render some small drawing.