Gpu: added rect texture flip x,y support
Drawing textured rectangle has some extra features than using Polygon,
and one of them is texture flipping, in which we can flip the texture by
the X or Y axis (or both).

This commit does not test flipping for CLUT textures yet (need to find a
test/game that does that).