Gpu: improved performance of vram writes
Using `write_at_position` was a big mistake, a lot of the time is spent
on mapping the buffer and then writing only one value. So now we only
write in blocks.

Also used the new way of converting u16 to pixel data using the new `ClientFormat`
that I introduced in a PR `https://github.com/glium/glium/pull/1943`.