Gpu: added texture support (Vulkan) and used manual swizzling
For the format I'm now using A1R5G5B5

The problem is that A1B5G5R5 (which is the PSX format) is not available
in Vulkan. Vulkan has component swizzling, but it only works for
texturing, and not writing or used as Framebuffer.

So in the end, the solution was (can be changed):
- When rendering (draw_polygon) write swap the blue and red.
  During this, the `render_image` in `gpu_context` will always have
  swapped components.
- To fix the swap, we will swap them again using Vulkan's swizzle
  thingy, but it doesn't seem to be an easy way to blit swizzle. So I
  made my own, which is just a simple render pipeline that takes the
  whole `render_image` as texture (and use swizzling on it) and then
  just render a whole rectangle to the front image.

This solution is really hacky and want to remove or at least improve it.

I created an issue to Vulkan to support a new format: https://github.com/KhronosGroup/Vulkan-Docs/issues/1691