Gpu: added transparency support using a second render_image as buffer
I couldn't find a good way to achieve full transparency features with
blends, so I created a second image and used it as a texture.

There is a heavy performance take from memory copy on each draw. To
improve performance a bit, maybe we can use the blends for semi modes
0,1,2 and only copy and do this special thing in mode 3. But when we add
mask support, I think this is the best approach on the long run.