Cpu: improved performance of register reads/writes
Because we were using fancy enums with values inside some of them (e.g
V(0), V(1)...) it was using branches to write to them.

Instead now we use an array, and we write to it by an index, which is
linear and does not contain any branches.

We kept the previous enums for formatting and printing when needed.