Memory: remove forcing mem_ctrl1 registers to a fixed value
These worked until now, but the bios is setting another value to
`cdrom_delay`. So, we needed to remove the restriction.

For now we are storing these registers as values, as I have no idea
what they actually mean. In the future, if they cause issue, we will
implement them properly.