Cdrom + Dma: implemented ReadN, Pause cdrom commands and Cdrom DMA
The implementation is very bad now, we have to go from dma to cdrom for
each byte, when we can just copy from cdrom_data array directly into
ram. But, I want to use the same design as old dma channels (which is
also bad), and get cdrom working, then later we should rewrite the whole
dma mechanism.

Not sure how correct is the implementation for ReadN and Pause commands,
its not clear, how cdrom interrupts should be handled for consequent
reads, but I think this should work fine, and the bios seems to be
working fine with this now.