Cdrom: added delay when running commands
The BIOS does not expect an interrupt to happen immediately after
issuing a command (which what we were doing), because it sometimes issues
the command inside a "critical section" (irq handler) and then clears
the interrupts before exiting the "critical section".

The delay time is 255 clock, it works now, but because we are using fixed
value it can give us trouble in the future. Maybe we should think of a
better way to handle delays.