Cdrom: implemented data fifo buffer
The reads commands are buffered (up to unknown extent, I just didn't
check for any limit for now) and then when requested from the
`request_register` it should copy to the fifo buffer.

For now there is a lot of copying done here, maybe we can do better
by just saving references to one buffer.