Gpu: moved gpu rendering to another thread
This is just moving the previous implementation to another thread, there
is no improvement, so its still slow.

There is also a stupid case with `gpu_read`, I added a delay when
reading from the channel, since sometimes, the backend isn't fast enough
to send the whole thing faster than the dma.