Gpu: re-use same `CommandBufferBuilder` for as many commands as possible
This design now, will try to use the same buffer until any of the
following happens:
- front_blit
- drawing textured/transparent polygon/line
- read command

Because these require last state of the vram, the front_blit and reading
is ok, but would be nice to optimize for textured and transparent
drawing.

This design now, makes emulation faster for drawing many opaque
polygons/lines, which is a nice improvement.