Gpu: better `utilization` of a single `command_buffer_builder`
By copying to the back image before setting it as texture, it's not
triggering conflict error anymore (not sure if its a good thing or a bug
from vulkano).

But even with this, stacking many commands in one buffer is very bad.
The bottleneck happens in
`vulkano::command_buffer::synced::builder::append_command`, in which I
think its checking for any conflicts with ALL previous commands. Not
sure if it can be improved, but limiting the number of commands in a
single command buffer seems to help.