Improving performance by reducing the number of components clock calls
The PSX is not timing sensitive hardware (unlike the NES or GB), which
allows us to clock the components at a non regular intervals, since the
BIOS will wait anyway for specific events/interrupts/flags to be set
before performing further actions.

This allows for very fast performance (especially by reducing the number
of times the `GPU` clock is being called, since it was being called a
lot just to increment a counter by `1`).

Right now, I think the DMA is the bottleneck, I'm running the DMA way
faster than it should (for example, DMA GPU channel transfer rate is 1
clk/word, but I'm transferring the whole thing in 1 cpu clock).