Dma: handled DMA with one call per `clock_components`
Similar to what we have done to the other components, now for the DMA.
Sometimes, `cycles` can be high, in that case, we transfer multiple
times to use all the `cycles` we get.

With this, a major performance boost is added, in the main thread now
around 90% of the execution time is done in the `CPU` alone (tested with
crash2), this means that the other components are very light to clock
(GPU being in another thread).

The next step would be to try to optimize the CPU somehow :)