Cpu: checking for `irq` once, and using `Instruction` reference
When looping from outside, we will check for `irq` and other checks
every time, which is not needed, since the bus is not clocked currently,
and will only clock after we return.
I'm not sure if `Instruction` will be passed by reference during
optimization if its not edited or copied, but its better to be explicit
about it.