Dma: run the DMA alongside the CPU and not by using `cpu_cycles`
The DMA in design is a component that uses the CPU cycles, and the CPU
will be paused while the DMA is running (not sure about modern CPUs).
Thus, we cannot delay the DMA exection if it was triggered (some
games were crashing already).

And also because the DMA is running of the CPU clock, it should be
treated the same as the CPU, and use the DMA cycles to clock the other
components.
But, when we did that, sometimes the DMA will clock for too many cycles,
that's why we split the cycles if there are too many.