Gpu: removed `back_image_update` on textured draws
Updating the back image on every texture draw is very heavy and not
needed (unlike transparency). Removed updates on every draw now.

Currently, the back image will be updates if vram was written to
directly (which is mostly for textures), and also once we create a new
command buffer builder.
The only glitch that can happen is if the GPU draw some polygons then
uses these as texture within the MAX_BUFFERED_COMMANDS limit.