Gpu: used Vulkan's Blending for semi_transparency (except mode 3)
This way we minimize the number of `update_back_image`. Previously, I
was using Blending with OpenGL, but because mode 3 cannot be implemented
with it, I moved to `back_image`, but we can use both, and that's what
I did here.

If there is no semi_transparency, the mode would normally be `0`, but
for some reason I get artifacts when that happens. When running
the `transparency` gpu test, everything looks normal, but I get some
artifacts from Crash2. Setting it mode 3 fixes it because it doesn't
have any blending. Would need to look into it more though.

One of the reasons I was using `back_image` is to implement masking
later, I'm planning to use stencil to implement masking draws, but not
sure how to get the mask to be written to the alpha channel.