Gpu: Buffered as many polygon vertices as possible to reduce `CmdDraw`
We are now adding most of the `push_constants` data to each individual
vertex. This way, we can buffer many polygons that may have different
parameters.

We buffer as many vertices until any of these events happen:
- Change in `semi_transparency_mode` (because we need to use different
  pipeline).
- Change in drawing area coordinates (planning to change this so we
  don't use viewport) and drawing offset.
- Drawing Lines (currently lines are not buffered).
- End of frame (front blit).

This boosted the performance a lot, especially with frames that contain
many triangles.