MDEC + DMA: implemented color decoding and MDEC out DMA
Used the decoding functions and algorithms, but this does not work for
most games, since you can send multiple blocks and read them all at the
end.

But for now, we only decode one block. We should read the data as they
come from the input fifo and put the output in output fifo as they are
done being processed.