Front+Cpu: added toggle-able flag to trace instructions
Using `RUST_LOG` to trace the CPU is very annoying since it traces
everything from the boot, so its very hard to debug something in the
middle of the game.

This was added to fix this issue, and to allow debugging at any time.

The performance drops significantly XD, from 60 FPS to 1 or less.
If we added support for writing to file directly, it might improve the
performance.