Cpu: added debug commands to set `write` breakpoint and view memory
The stack is a bit useless, I thought maybe we can build the stacktrace,
but it appears to be hard or impossible, a better approach would be to
add address to an internal structure on `jal/jalr/...` instructions.

Also, the write breakpoint has a bug where the PC is wrong on printing,
since when we are executing, the `regs.pc` is already advanced.