Gpu: added 24bit direct rendering
This is done by reading the vram buffer as 24bit colors instead of
16bit. In order to do that, we used compute shader to extract each color
and write them into a new image for display.

This is not very efficient because we have to extract each byte
individually. Using compute shader helps a lot in performance, but we
have to optimize it a bit more.

On thing we can do is to have a totally separate code block to handle
24bit into the `dest_image` directly, without having to use a middle
image.

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>