Debugger: fixed cmd history, and kept it available between pauses
When using the step command, the emulator is being not paused for one
instruction, because we are creating a new editor on every pause and
resume, we are losing the old history.

Thankfully, since Rustyline provide `&mut History`, we can replace but
with an already saved session, and take it when resuming.