Cdrom: fixed "not interrupting" bug in ReadN/S cmd
When there is data in cdrom, and the user doesn't want data now, it will
reset the interrupt register and not send "want data". Before this fix,
we only send interrupt once, and will only send it again when there is
new data.
But since the user will want to get interrupt always, then we should
send interrupts even if the user doesn't want it now, until they do.