Gpu: added fast path for semi-transparency mode 3 non-textured
Mode 3 transparency is the bottleneck of GPU performance, it was running
very badly in Crash2 on rainy levels, since it uses semi-transparency 3
to render rain lines.

This addition, makes the level runs normally, but in case if a game uses
semi-transparency 3 with textures, we will return to the same issue.

We need to find a better way to handle this.