Spu: Fixed bug in sample indexing resulting in noise
Because we relied on index=0 to be a new state and a finish of a block,
there was an issue in indexing.

At first, I thought index will always increment, and thus index=0 will
only be present on the first sample only, but if the rate is less than
44100 (i.e. pitch parameter < 0x1000), then the index won't increment
and stays at 0, which at the next clock will fetch a new block and not
use the current new block, and so on.