Cpu: changed instructions/cycles measurement (Faster performance ...
and fixed Vsync timeout)

The reason for low FPS before, is that we were executing too many CPU
cycles per frame, and when we used the correct GPU clock, it resulted in
Vsync timeout issues.
This is mostly because the timeout is calculated by approx number of instructions?

Anyway, now we can run 60FPS smoothly.

The current cycles are not accurate at all, since I couldn't find a
documentation for number of cycles for each instruction, but this work fine
as the CPU doesn't need to be cycle accurate.