Controller: added config mode and some config commands
The controller contain config mode, which behave differently than normal
mode, and many games are using it.
Right now, it has implementation of what seems to be needed, there are
other config commands, but the remaining ones are all `Unknown`
according to the docs, so will add them if they are needed.