Gpu: added support for wrapping vram_read/write
Looks like the PSX gpu performs wrapping when accessing out-of-range
bounds.

The current process of doing this is to create a temporary image
buffer to copy from/into in case of any overflow, and handle all edges.
Then use this temp image to copy to buffer (vram_read), or read from buffer
(vram_write) in the beginning.