Gpu: improved performance by reducing `update_back_image` if not needed
When there are multiple writes to the vram, its not needed to update the
back image every time, instead, we can just update before draw if we
need it.

Do note that with the current design (even before this commit), if you
draw something and used it as a texture, it might not work if you didn't
issue any command that would trigger `back_image` update.