Gpu: moved most gpu_stat and flag processing into cpu thread
Now, only rendering commands are sent to the backend.

This actually generates another issue for us, because we are using Mutex
for the shared state, it will be out-of-sync almost all the time when
the backend executes the command, it will contain the same data that was
there when the command first started.