Gpu: send current gpu_state to backend on command exec instead of atomic
Using atomic is not accurate, as the command executing on a different
state than was when the command was first issued.

With this design, there is not need to have Mutex to keep lock, since
its only modified on a single thread, and a new copy is sent with every
command that needs it.