Gpu: added handling for interrupting running CpuToVramBlit command
If GP1(01) was executed, it should flush the command buffer (I assume
also reset the current command), for anything else, it should be ok,
since it won't be sent to the backend anyway, but for CpuToVramBlit we
want to sent the data that have been given from the cpu.

Currently, if CpuToVramBlit was interrupted, it will only send full rows
(unless the first row is not even filled). That's because in order to
send half rows, we would need to send multiple `VramWriteBlock` backend
command, which currently we can't do.