Gpu: fixed texture misbehavior for small textures
This affected fonts and some textures in some games (like the characters
in FinalFantasy I).

This was due to using `round` when computing the texture position, one
fix is to replace `round` with `floor` (works better than ceil), but
here, we just used float everywhere except for getting CLUTs, and it
works very well.

I thought that using `round` would be similar behavior to what NEAREST
filter is doing, but looks like it was not.