Gpu: partially fix for texture_flip in rectangles
The previous implementation (and this) are very wrong, but this at least
is in the right direction.

The way flip works, is that it will decrement the counter for the
current texture pixel when rendering, since we don't have access to it,
we have to tell the gpu of the texture_coord, and it will do the
interpolation on its end (inc/dec based on other points).

The previous implementation performs flips for the whole tex_page, which
is not how this works. I was mislead, because I was testing with a test
that flips the whole page.