Front: Don't clock full-frame at once, and leave room for UI response
Previously, the UI would be very jagged when trying to resize or close
the window, because most of the time, the UI doesn't have control as we
are on the limit of the FPS (60). This allows the UI to get some room
for updates and handling other events, also makes input much more
responsive.