Cargo: update dependencies, especially `calloop >=0.10.4` for bug fix
There is a weird bug in calloop that would result in crashing, this is
the library used by `winit` to manage the events. I noticed this would
happen when clicking a lot of buttons at the same time. This is probably
some sort of race condition.

Anyway, it got fixed in `0.10.4`