Deps: updated `bitflags` to `2.1`, and fixed some issues with bitflags
A lot changed in this version. The major break is the behavior of
`from_bits_truncate` changed now, before it would do this
```
bits = input & Self::all()
```
This mean, if you have this example
```
Flags {
    A = 0b0011
    B = 0b0100
}
```

And for example you input `0b0110`, you will get this in the `bits` as
0b0110 & 0b0111 == 0b0110`. But now it changed and `truncate` now will
only allow full representation of the flags, so for the above input the
`bits` will be `0b0100` as `0b10` doesn't match `0b11` fully.

But as a fix, They added another function `from_bits_retain`, but this
is also not perfect as it just copies the input as is, without doing any
checking, which would mean the input `0b1111` would be set as is, but it
should be `0b0111` instead.

Anyway, we are using `retain` anyway, it doesn't seem to be causing
issues for now, but I suspect it would result in failure in some
hardware tests where they expect some bits to be ignored by hardware.

The thing is, I was always not sure of using bitflags for what we have
been doing, it doesn't seem to be designed for our use case where
several bits can be in any arrangement. TODO: look/create a better
library that handles these kind of hardware flags/bits.