Cpu+Debugger: added Instruction parsing to the debugger and public
Not the best way, we made a new `Instruction` struct that is using
`RegisterType` instead of `Register`, because `Register` is much easier
to construct and faster, but is not intuitive to use for outside, as it is just an
integer.

This works fine, but would be nice to get a better solution to getting
access to `Instruction`.