Cpu: reduced the parsing on `Instruction` for performance
We are still parsing the `rd/rs/rt` raw values, but there is no need to
convert them to `RegisterType` which is quite heavy.

The imm values could have been also included in the parsing, but I think
they are used a lot less than the alternatives, so they are lazily
parsed as well.