Cpu+Debugger: added `step-out` command to step out of a function
The way this is implemented is very simple and can be not accurate in
some cases. It just sets breakpoint on the last element on the call
stack.
This can cause problems because the call stack is not accurate,
especially when dealing with kernel functions like A0/B0/C0, which are
syscall dispatches.

If we get a reproducible bug, I think it can be fixed, so I'll leave it
to that time.

Another note about this, is that we are adding another HashSet with
another search on every instruction, which is quite heavy.
An idea I have is to create a `HashMap` for all types of breakpoints and
then we will only have 1 lookup for the current address to check for all
types of breakpoints associated with it.