Spu+Dma: added `DmaRead` for the spu and better handling for spu dma
Now the DMA has "direct" access to the spu as it should be. The flags
right now are a bit funky, we are setting the `BUSY` flag in the spu for
the whole time until the user stops it themselves, which I think is
wrong.

Not sure if any game/rom checks for this, but, for now this should work.