Front+Misc: added `fast_boot` and handled loaded EXE better
This method is taken from `pcsx-redux`, which just hijacks the execution
when arriving on the shell after everything is done setup and just skips
it.

The shell's purpose is to load the animation, check the cdrom is
available and verify it, if there is no disk, it will continue to open
the typical PSX UI which appear when nothing is inserted.
If there is a disk, it will just return, and the caller will load the
disk and pass execution to it I assume.