Cdrom: implemented buffer overrun and adjusted Read delays
Sometimes, the read was too fast and the game would get stuck waiting
for the data because we hit interrupt while it was processing the
previous data. (happened in NSF2).

Looks like delaying at a specified interval is much better, and I guess
its better for the audio? since it must be in a specified frequency.