Cdrom: changed how we handle `read` to be a specific state instead of command
What this means, is that the read command only switch the cdrom to the
read state, and while its in the read state it will read sectors, this
means this will allow the cdrom to execute other commands while the read
is executing.

I think this is correct since this fixes an issue with
MegaMan8 intro not playing, since it was doing `Read` and them after
that `GetLoc`, which interrupts the `Read` command.