Dma: fix priority handling of concurrent DMA channels
This fixes an old issue where on MDEC video handling, we are seeing some
artifacts in the middle of the frame.

The reason for this is that, STR video decoding works by submitting a
full frame of MDEC compressed data into the MDEC by using `MDEC_in DMA`
channel for instance, and then repeatedly requesting a slice from
`MDEC_out DMA` and feeding this slice into `GPU DMA` by using something
like `LoadImage` function.

Now, our priority was like this:
- highest priority: MDEC_in
                    MDEC_out
                    GPU
                    CDROM
                    SPU
                    PIO
- lowest priority:  OTC

How exactly the synchronization is being done is like this:
- Implement a function as DecDCTOut callback handler `handler`
- Set DecDCToutCallback to `handler`, this function will be called once
  `MDEC_out` DMA is finished.
- Inside the function, you will take the data and feed it into
  `LoadImage`, switch to the next buffer, and start `MDEC_out` again for
  the next buffer, since mostly double buffered is used.

Now, the issue happens when you get a callback that `image1` is done, so
you start 2 DMA transfers, one to move `image1` to the GPU and another
to start filling `image2` with the next slice.

Because of our wrong priority handling, filling `image2` will finish
before `image1` is finished moving to the GPU. Then we will start
another data filling for `image1`, since the game "assumes" its free
now.

Fixing the priority, will first move `image1` to the GPU before it does
anything with `image2`.