Front: improved FPS handling for higher rate monitors
The issue was that the monitor would refresh at a rate not multiple of 60, like 144, which results in events only appearing at 144 rate.
This resulted in the FPS becoming 58~ because the elapsed would be something like
```
15000
16055
17111
// resets here
```
As you can see, the place we want is in the middle between 16000 and 17000.
So, implemented a way to record these differences in an approximate way to fix this issue.

Also used `FPS` struct to do it since we were keeping track of the current time in multiple places.