Gpu: don't perform `VramVramBlit` if not needed
Some games (Need for speed High Stakes) does a very minimal VramVramBlit.
`VramVramBlit: (0..2, 0..1) -> (0..2, 0..1)`
Its very small, but costs a lot, because we need to get the data and flush all the buffers.

If we detect this kind of blit, we shall ignore.

In the future, we should consider using vulkan APIs to perform in-gpu copy.