Spu: Fix bug where `adpcm_repeat_address` is being set wrongly
This result in an offset of `2`, which is very important especially if a
game relies on interrupt being triggered from these missed addresses.

This game was `Need for speed - II`, it won't go through the main screen
and the audio was looping (because it didn't receive message that it can
go on to the next audio block.

Signed-off-by: Amjad Alsharafi <26300843+Amjad50@users.noreply.github.com>